fix(clipboard): cancel "Changed" emissions on destruction
The clipboard uses an idle source to defer emission of the `Changed`
signal, to avoid concurrency issues when reading the clipboard.

Ensure that the idle source is removed upon destruction, so that we
don't call back to a finalized object. Reuse the variable used as a
rentrancy guard to hold the source ID.